Shorten number feature

+++ Vampirism Idle/Assets/Scripts/BloodGatherManager.cs
@@ -68,7 +68,7 @@ public class BloodGatherManager : MonoBehaviour //Persistent
                 GameManager.res_HumanPop -= VampireManager.vampires_amount_Used_Infect[vampIndex]; //decrease pop
 
                 /* Infection Code */
-                vampiresToAdd = Mathf.FloorToInt(VampireManager.vampires_amount_Used_Infect[vampIndex] * (float)(VampireManager.vampires_InfectionChanceTotal[vampIndex] / 100f)); //calculate amount of vampires to add; 10% of vamps will convert a human
+                vampiresToAdd = Mathf.FloorToInt((float)VampireManager.vampires_amount_Used_Infect[vampIndex] * (float)(VampireManager.vampires_InfectionChanceTotal[vampIndex] / 100f)); //calculate amount of vampires to add; 10% of vamps will convert a human
 
                 //If housing is greater than the vampires to be added
                 if (HousingManager.housing_amountAvailable > vampiresToAdd)
@@ -85,7 +85,7 @@ public class BloodGatherManager : MonoBehaviour //Persistent
             {
                 GameManager.res_Blood += VampireManager.vampires_bloodPerInfect[vampIndex] * GameManager.res_HumanPop;
 
-                vampiresToAdd = Mathf.FloorToInt(VampireManager.vampires_amount_Used_Infect[vampIndex] * (float)(VampireManager.vampires_InfectionChanceTotal[vampIndex] / 100f));
+                vampiresToAdd = Mathf.FloorToInt((float)VampireManager.vampires_amount_Used_Infect[vampIndex] * (float)(VampireManager.vampires_InfectionChanceTotal[vampIndex] / 100f));
 
                 //If vampires to be added is greater than the pop, convert all humans to vamps
                 if (vampiresToAdd > GameManager.res_HumanPop)
